backend/image-resizer/requirements.txt
pika==1.3.2
Pillow==10.4.0
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9
pytest==8.2.0